# Conflicts:
#	sbeam/target/classes/com/sbeam/sbeam/config/XxlJobConfig.class
#	sbeam/target/classes/com/sbeam/sbeam/controller/PaymentRecordsController.class
#	sbeam/target/classes/com/sbeam/sbeam/entity/Game.class
#	sbeam/target/classes/com/sbeam/sbeam/job/OrderToHistoryJob.class
#	sbeam/target/classes/com/sbeam/sbeam/mapper/GameMapper.class
#	sbeam/target/classes/com/sbeam/sbeam/service/IPaymentRecordsService.class
#	sbeam/target/classes/com/sbeam/sbeam/service/impl/PaymentRecordsServiceImpl.class
#	sbeam/target/classes/com/sbeam/sbeam/util/RedisLuaExecutor.class
#	sbeam/target/test-classes/com/sbeam/sbeam/SbeamApplicationTests.class
#	sbeam/target/test-classes/com/sbeam/sbeam/TestMybatisGenerate.class

+++ sbeam/src/main/java/com/sbeam/sbeam/service/IUserService.java
@@ -15,7 +15,7 @@ public interface IUserService extends IService<User> {
 
     User createUser(User user);
 
-    User getUserById(Long id);
+    User getUserById(Integer id);
